requirements.txt
Flask>=2.0,<3.0
RPi.GPIO>=0.7
waitress>=2.1
